+++ repro/models/susanto2016.py
@@ -1,0 +1,2 @@
+from ._models.susanto2016.src.model import RNNTruecaser
+from ._models.susanto2016.src.setup import Susanto2016SetupSubcommand